requirements.txt
discord.py>=2.1
python-dotenv>=0.20.0
colorama>=0.4.6
rich>=13.0.0